删除冗余代码
